A few very minor changes
Made a lot of these on my own bot ages ago but never updated the project
so here they are:

- Updated whisper listener to work with tmi.js changes.
- Updated tmi.js settings from "random" to "cluster" and changed the
chat cluster to AWS.
- Bot will not get stuck if a channel has banned the hosting bot and it
tries to host on their channel.
- The "getTeamLiveChannels" will now always return an array, even if it
fails to access the Twitch API. In this case, the array will be empty so
the bot will assume no one is online and check again later.
- Added "turnOnHosting" callback function to the start.js script because
somehow I forgot to before.