Rewriting some double host prevention
If the channel goes offline but the API still reports them as online
(can happen in some cases) this will prevent them from being hosted
again. I already had protection in for this but it caused some other
issues. This still isn't perfect; I might try to make it a bit cleaner.