Added double check for if target channel offline
Similar to the other issue, sometimes it would detect the target channel
as offline if they weren't, and sometimes they were the only channel
online so it would just rehost them 5 minutes later. This now will only
happen if it's reported as offline 2 times in a row with a gap of 5
minutes between the 2 checks.