Small fix for getTeamChannels/getTeamLiveChannels
Sometimes when we request the team webpage, a page can show someone as
online, but when the next page is loaded, they are offline. This usually
means you get a duplicate user in the list and someone is seen as
"removed" from the team. If this happens now, it will restart the check
from page 1.