Adding the ability to have a "custom" team
If the user defines an array of channels in the config using the
"teamChannels" variable, it will use that list of channels instead of
querying Twitch for channels on that team. The team name will then
become "custom-{name}" within the program/code itself, so the "team"
variable can be whatever.