Minor fixes
Mainly a minor fix to the hosting code I just change and also made it so
the program doesn't incorrectly report a new version is available when
the running version is actually higher than the "current" version. Also
fixed it so that new logs do not start with "undefined" which broke the
webpage JS a bit.