Changed the way channel online status is monitored
The code now uses Twitch's new PubSub service to keep an eye on all
channels for when they go online/offline and periodical viewer count
updates. The API still has to be queried every 5 minutes per online
channel to check what their currently set game is (as some parts of the
code need it). We now keep a object containing all online channels of
the ones that we are monitoring, and this is used to check for people to
host and channels to host on.