Added monitoring for online channels if no one live
If no one is live when the bot attempts to host someone, instead of
checking every 5 minutes, it will now listen for an event to say if
someone on the team has gone live, and then attempt to host again. This
code could use some work, but it's in testing for now.