Slight change to fix from earlier for monitoring
My fix from earlier caused a new problem: if we were monitoring for
online channels while someone was stilling online but had gone over the
normal hosting time, and then they went offline, the online monitoring
was cleared so the bot got stuck. I have gone back to the old way now
but if a manual host is triggered, online monitoring is stopped if it's
currently happening.