Added feature to stop our host replacing other hosts
Before, if a channel was being hosted, then they hosted someone else and
then went offline, the bot would then replace their host with it's own,
which some people didn't like. Now if that happens they are added to a
list of channels the bot will not attempt to host on and then are
removed after 25 minutes.