Fallbacks for if statistics.json doesn't load correctly
Recently I've had issues with the statistics.json file not loading in
correctly so all data is lost. This code brings in backups, 1 backup for
each time the application is started, and if the file fails to load we
now try 4 times to make sure. Some of this is a bit overkill, it will
probably be removed in the future if I know when/why the bug happens.