Slight improvement for repeated code
Wrote a function to return if we are currently hosting a channel or if
we were last hosting a channel. Also added it somewhere else to fix an
issue of when a channel is being hosted and we are currently monitoring
for channels to come online, and then the hosted channel goes
offline/online themselves in a short timespan.